Add the Push Gateway URL to the trial job
